requirements.txt
streamlit
pandas
openpyxl
pypdf
google-generativeai